fix: isolate OpenCode provider protocols
