+++ app/src/main/java/com/ai/assistance/operit/api/chat/llmprovider/GeminiProvider.kt
@@ -82,7 +82,7 @@ internal data class GeminiThinkingConfig(
 }
 
 /** Google Gemini API的实现 支持标准Gemini接口流式传输 */
-class GeminiProvider(
+open class GeminiProvider(
     private val apiEndpoint: String,
     private val apiKeyProvider: ApiKeyProvider,
     private val modelName: String,
@@ -1242,8 +1242,6 @@ class GeminiProvider(
             preserveThinkInHistory: Boolean = false
     ): RequestBody {
         val json = JSONObject()
-        val isOpenCode = OpenCodeReasoningParameters.isMarked(modelParameters)
-
         // 添加工具定义
         val tools = JSONArray()
         
@@ -1287,7 +1285,7 @@ class GeminiProvider(
 
         // 添加模型参数
         for (param in modelParameters) {
-            if (param.isEnabled && !OpenCodeReasoningParameters.isInternal(param)) {
+            if (param.isEnabled) {
                 when (param.apiName) {
                     "temperature" ->
                             generationConfig.put(
@@ -1337,7 +1335,7 @@ class GeminiProvider(
             }
         }
 
-        if (enableThinking && !isOpenCode) {
+        if (enableThinking) {
             val thinkingQualityLevel =
                 runBlocking {
                     ApiPreferences.getInstance(context).thinkingQualityLevelFlow.first()
@@ -1364,21 +1362,16 @@ class GeminiProvider(
     }
 
     /** 创建HTTP请求 */
-    private suspend fun createRequest(
+    protected open suspend fun createRequest(
             context: Context,
             requestBody: RequestBody,
             isStreaming: Boolean,
             requestId: String
     ): Request {
-        // OpenCode already provides a model-specific endpoint; Google endpoints use the standard path.
+        // 确定请求URL
+        val baseUrl = determineBaseUrl(apiEndpoint)
         val method = if (isStreaming) "streamGenerateContent" else "generateContent"
-        val requestUrl =
-            if (apiEndpoint.contains("/models/", ignoreCase = true)) {
-                "${apiEndpoint.trim().removeSuffix("/")}:$method"
-            } else {
-                val baseUrl = determineBaseUrl(apiEndpoint)
-                "$baseUrl/v1beta/models/$modelName:$method"
-            }
+        val requestUrl = "$baseUrl/v1beta/models/$modelName:$method"
 
         AppLogger.d(TAG, "请求URL: $requestUrl")
 
@@ -1392,23 +1385,17 @@ class GeminiProvider(
 
         // 添加API密钥
         val currentApiKey = apiKeyProvider.getApiKey()
-        val requestBuilder = builder
-            .url(requestUrl)
-            .post(requestBody)
-            .addHeader("Content-Type", "application/json")
-        if (apiEndpoint.contains("/models/", ignoreCase = true)) {
-            requestBuilder.addHeader("Authorization", "Bearer $currentApiKey")
-        } else {
-            requestBuilder.url(
-                if (requestUrl.contains("?")) {
-                    "$requestUrl&key=$currentApiKey"
-                } else {
-                    "$requestUrl?key=$currentApiKey"
-                }
-            )
-        }
+        val finalUrl =
+            if (requestUrl.contains("?")) {
+                "$requestUrl&key=$currentApiKey"
+            } else {
+                "$requestUrl?key=$currentApiKey"
+            }
 
-        val request = requestBuilder.build()
+        val request = builder.url(finalUrl)
+                .post(requestBody)
+                .addHeader("Content-Type", "application/json")
+                .build()
 
         logLargeString(TAG, context.getString(R.string.gemini_request_headers, HttpLogSanitizer.headersForLog(request.headers)))
         return request
