+++ docs/TODO/opencode_provider_isolation_20260820/03-verification.md
@@ -1,0 +1,3 @@
+# 验证记录
+
+本次按仓库执行准则不运行构建、Gradle 或测试命令。完成后执行静态 diff 检查，并核对公共 provider 不再引用 OpenCode 专用类型。
